add: qrcode login.
